Joc penjat: canvis de estructura i encapsulació. + llista de lletres fallades.
